удалил компонент FilmInfo, вместо него Tabs
